Add files via upload
thank for comments gilgil mentor !!!
now i could unserstand pointer & structure pointer because of his task~

+++ pcap_test.h
@@ -2,8 +2,8 @@
 #define PCAP_TEST_H
 #include <stdint.h>
 #pragma once
-void print_port(uint8_t *port);
-void print_ip(uint8_t *ip);
+void print_port(uint16_t *port);
+void print_ip(uint32_t *ip);
 void print_mac(uint8_t *mac);
 
 #endif // PCAP_TEST_H
